Improved (Z, N) to inde conversion
After (very roughly) profiling the code, it seems that a huge portion of the network solving code is spent accessing the `zn_to_index_dict` dictionary. It appears that having a [Z, N] vector as a key is very inefficient. This commit introduces a different way of accesign this index information from the [Z, N] pair.

This commit also adds a new struct, `NetworkIndex`, which combines all of the information one needs when converting from [Z, N] to index. It also holds the `mass_vector` variable which holds the A=Z+N values for a given index.

To convert from [Z, N] to index the function `zn_to_index` has been added.

To check if a species [Z, N] is in the range of the network, the function `zn_in_network` has been added.

Documentation will be added in a future commit.