Added the ability to modulate the rates of reactions by arbitrary muliplicative factors, as specified in the control file.
Added the ability to read in .dat files (as well as .jld files). Reading in .dat files is slower than .jld files, but could be more convienient for the user.

Fixed an issue with reading in data that caused the runtime to be quadratic in the number of reactions to read. The algorithm is now linear in time by using dictionaries to store ReactionDataIO members.